Refactored people routes into own file. Fixed some bugs with app.use ordering.
